@@ -1,11 +1,18 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace ASP.NET.Kitchen
+using System.ComponentModel.DataAnnotations;
+
+namespace Kitchen;
+
+public class PrepareDish
 {
-    public class PrepareDish
-    {
-        public string Chef { get; set; } = "";
-        public string Dish { get; set; } = "";
-    }
+    [Required]
+    public string Chef { get; set; } = "";
+    
+    [Required]
+    public string Dish { get; set; } = "";
+    
+    [Required]
+    public string Kitchen { get; set; } = "";
 }
