@@ -6,14 +6,12 @@ using System;
 using Dolittle.SDK.Events;
 using Dolittle.SDK.Events.Handling;
 
-namespace Kitchen
+[EventHandler("f2d366cf-c00a-4479-acc4-851e04b6fbba")]
+public class DishHandler
 {
-    [EventHandler("f2d366cf-c00a-4479-acc4-851e04b6fbba")]
-    public class DishHandler
+    public void Handle(DishPrepared @event, EventContext eventContext)
     {
-        public void Handle(DishPrepared @event, EventContext eventContext)
-        {
-            Console.WriteLine($"{@event.Chef} has prepared {@event.Dish}. Yummm!");
-        }
+        Console.WriteLine($"{@event.Chef} has prepared {@event.Dish}. Yummm!");
     }
 }
+
