@@ -6,26 +6,24 @@ using System;
 using Dolittle.SDK.Aggregates;
 using Dolittle.SDK.Events;
 
-namespace Kitchen
+[AggregateRoot("01ad9a9f-711f-47a8-8549-43320f782a1e")]
+public class Kitchen : AggregateRoot
 {
-    [AggregateRoot("01ad9a9f-711f-47a8-8549-43320f782a1e")]
-    public class Kitchen : AggregateRoot
+    int _ingredients = 2;
+
+    public Kitchen(EventSourceId eventSource)
+        : base(eventSource)
     {
-        int _ingredients = 2;
-
-        public Kitchen(EventSourceId eventSource)
-            : base(eventSource)
-        {
-        }
-
-        public void PrepareDish(string dish, string chef)
-        {
-            if (_ingredients <= 0) throw new Exception("We have run out of ingredients, sorry!");
-            Apply(new DishPrepared(dish, chef));
-            Console.WriteLine($"Kitchen {EventSourceId} prepared a {dish}, there are {_ingredients} ingredients left.");
-        }
-
-        void On(DishPrepared @event)
-            => _ingredients--;
     }
+
+    public void PrepareDish(string dish, string chef)
+    {
+        if (_ingredients <= 0) throw new Exception("We have run out of ingredients, sorry!");
+        Apply(new DishPrepared(dish, chef));
+        Console.WriteLine($"Kitchen {EventSourceId} prepared a {dish}, there are {_ingredients} ingredients left.");
+    }
+
+    void On(DishPrepared @event)
+        => _ingredients--;
 }
+
