@@ -3,20 +3,18 @@
 
 using Dolittle.SDK.Events;
 
-namespace Kitchen
+[EventType("b27f2a39-a2d4-43a7-9952-62e39cbc7ebc")]
+public class EmployeeTransferred
 {
-    [EventType("b27f2a39-a2d4-43a7-9952-62e39cbc7ebc")]
-    public class EmployeeTransferred
-    {
-        public string Name { get; set; }
-        public string From { get; set; }
-        public string To { get; set; }
+    public string Name { get; set; }
+    public string From { get; set; }
+    public string To { get; set; }
 
-        public EmployeeTransferred(string name, string from, string to)
-        {
-            Name = name;
-            From = from;
-            To = to;
-        }
+    public EmployeeTransferred(string name, string from, string to)
+    {
+        Name = name;
+        From = from;
+        To = to;
     }
 }
+
