@@ -2,35 +2,23 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // Sample code for the tutorial at https://dolittle.io/tutorials/getting-started/csharp/
 
-using System.Threading.Tasks;
 using Dolittle.SDK;
 using Dolittle.SDK.Tenancy;
+using Microsoft.Extensions.Hosting;
 
-namespace Kitchen
-{
-    class Program
-    {
-        public static async Task Main()
-        {
-            var client = DolittleClient
-                .ForMicroservice("f39b1f61-d360-4675-b859-53c05c87c0e6")
-                .WithEventTypes(eventTypes =>
-                    eventTypes.Register<DishPrepared>())
-                .WithEventHandlers(builder =>
-                    builder.RegisterEventHandler<DishHandler>())
-                .Build();
+var host = Host.CreateDefaultBuilder()
+    .UseDolittle(clientBuilder => clientBuilder
+        .WithEventTypes(eventTypes => eventTypes.Register<DishPrepared>())
+        .WithEventHandlers(eventHandlers => eventHandlers.RegisterEventHandler<DishHandler>()))
+    .Build();
 
-            var preparedTaco = new DishPrepared("Bean Blaster Taco", "Mr. Taco");
+await host.StartAsync();
 
-            await client.EventStore
-                .ForTenant(TenantId.Development)
-                .Commit(eventsBuilder =>
-                    eventsBuilder
-                        .CreateEvent(preparedTaco)
-                        .FromEventSource("Dolittle Tacos"));
+var client = await host.GetDolittleClient();
+await client.EventStore
+    .ForTenant(TenantId.Development)
+    .Commit(eventsBuilder => eventsBuilder
+        .CreateEvent(new DishPrepared("Bean Blaster Taco", "Mr. Taco"))
+        .FromEventSource("Dolittle Tacos"));
 
-            // Blocks until the EventHandlers are finished, i.e. forever
-            client.Start().Wait();
-        }
-    }
-}
+await host.WaitForShutdownAsync();
