@@ -4,17 +4,14 @@
 
 using Dolittle.SDK.Projections;
 
-namespace Kitchen
+[Projection("98f9db66-b6ca-4e5f-9fc3-638626c9ecfa")]
+public class DishCounter
 {
-    [Projection("98f9db66-b6ca-4e5f-9fc3-638626c9ecfa")]
-    public class DishCounter
-    {
-        public int NumberOfTimesPrepared = 0;
+    public int NumberOfTimesPrepared = 0;
 
-        [KeyFromProperty("Dish")]
-        public void On(DishPrepared @event, ProjectionContext context)
-        {
-            NumberOfTimesPrepared ++;
-        }
+    [KeyFromProperty("Dish")]
+    public void On(DishPrepared @event, ProjectionContext context)
+    {
+        NumberOfTimesPrepared ++;
     }
 }
