@@ -1,28 +1,24 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using Dolittle.SDK.Events;
 
 namespace Dolittle.SDK.Aggregates
 {
     /// <summary>
-    /// Defines a way to work with <see cref="AggregateRoot"/>.
+    /// Defines a system that knows how to get the <see cref="IAggregateRootOperations{TAggregate}"/> for a specific <see cref="AggregateRoot"/>
+    /// scoped to a tenant.
     /// </summary>
-    /// <typeparam name="TAggregate"><see cref="AggregateRoot"/> type.</typeparam>
-    public interface IAggregateOf<TAggregate>
-        where TAggregate : AggregateRoot
+    /// <typeparam name="TAggregateRoot">The <see cref="Type"/> of the <see cref="AggregateRoot"/> class.</typeparam>
+    public interface IAggregateOf<TAggregateRoot>
+        where TAggregateRoot : AggregateRoot
     {
         /// <summary>
-        /// Create a new <see cref="AggregateRoot"/> with a random <see cref="EventSourceId"/>.
+        /// Gets the <see cref="IAggregateRootOperations{TAggregate}"/> for the <typeparamref name="TAggregateRoot"/> <see cref="AggregateRoot"/> class.
         /// </summary>
-        /// <returns><see cref="IAggregateRootOperations{T}">Operations</see> that can be performed.</returns>
-        IAggregateRootOperations<TAggregate> Create();
-
-        /// <summary>
-        /// Gets an<see cref="AggregateRoot"/> with a given <see cref="EventSourceId"/>.
-        /// </summary>
-        /// <param name="eventSourceId"><see cref="EventSourceId"/> of the <see cref="AggregateRoot"/>.</param>
-        /// <returns><see cref="IAggregateRootOperations{T}">Operations</see> that can be performed.</returns>
-        IAggregateRootOperations<TAggregate> Get(EventSourceId eventSourceId);
+        /// <param name="eventSourceId">The <see cref="EventSourceId"/> of the aggregate to get.</param>
+        /// <returns>The <see cref="IAggregateRootOperations{TAggregate}"/> for the <typeparamref name="TAggregateRoot"/> <see cref="AggregateRoot"/> class.</returns>
+        IAggregateRootOperations<TAggregateRoot> Get(EventSourceId eventSourceId);
     }
 }
