@@ -1,6 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
 using System.Threading;
 using Dolittle.SDK.DependencyInversion;
@@ -82,9 +83,11 @@ namespace Dolittle.SDK.Events.Handling.Builder
             IEventProcessors eventProcessors,
             IEventTypes eventTypes,
             IEventProcessingConverter processingConverter,
-            IContainer container,
+            TenantScopedProvidersBuilder tenantScopedProvidersBuilder,
+            Func<ITenantScopedProviders> tenantScopedProvidersFactory,
             ILoggerFactory loggerFactory,
-            CancellationToken cancellation)
+            CancellationToken cancelConnectToken,
+            CancellationToken stopProcessingToken)
         {
             var eventTypesToMethods = new Dictionary<EventType, IEventHandlerMethod>();
             if (!_methodsBuilder.TryAddEventHandlerMethods(eventTypes, eventTypesToMethods, loggerFactory.CreateLogger<EventHandlerMethodsBuilder>()))
@@ -114,7 +117,8 @@ namespace Dolittle.SDK.Events.Handling.Builder
             eventProcessors.Register(
                 eventHandlerProcessor,
                 new EventHandlerProtocol(),
-                cancellation);
+                cancelConnectToken,
+                stopProcessingToken);
         }
     }
 }
