@@ -41,7 +41,8 @@ namespace Dolittle.SDK.Events.Processing
         public void Register<TIdentifier, TClientMessage, TServerMessage, TRegisterArguments, TRegisterResponse, TRequest, TResponse>(
             IEventProcessor<TIdentifier, TRegisterArguments, TRequest, TResponse> eventProcessor,
             IAmAReverseCallProtocol<TClientMessage, TServerMessage, TRegisterArguments, TRegisterResponse, TRequest, TResponse> protocol,
-            CancellationToken cancellationToken)
+            CancellationToken cancelConnectToken,
+            CancellationToken stopProcessorsToken)
             where TIdentifier : ConceptAs<Guid>
             where TClientMessage : class, IMessage
             where TServerMessage : class, IMessage
@@ -50,14 +51,15 @@ namespace Dolittle.SDK.Events.Processing
             where TRequest : class
             where TResponse : class
         {
-            var processor = Task.Run(() => RunProcessorForeverUntilCancelled(eventProcessor, protocol, cancellationToken), cancellationToken);
+            var processor = Task.Run(() => RunProcessorForeverUntilCancelled(eventProcessor, protocol, cancelConnectToken, stopProcessorsToken), stopProcessorsToken);
             _processingCoordinator.RegisterProcessor(processor);
         }
 
         async Task RunProcessorForeverUntilCancelled<TIdentifier, TClientMessage, TServerMessage, TRegisterArguments, TRegisterResponse, TRequest, TResponse>(
             IEventProcessor<TIdentifier, TRegisterArguments, TRequest, TResponse> eventProcessor,
             IAmAReverseCallProtocol<TClientMessage, TServerMessage, TRegisterArguments, TRegisterResponse, TRequest, TResponse> protocol,
-            CancellationToken cancellationToken)
+            CancellationToken cancelConnectToken,
+            CancellationToken stopProcessorsToken)
             where TIdentifier : ConceptAs<Guid>
             where TClientMessage : class, IMessage
             where TServerMessage : class, IMessage
@@ -66,13 +68,13 @@ namespace Dolittle.SDK.Events.Processing
             where TRequest : class
             where TResponse : class
         {
-            while (!cancellationToken.IsCancellationRequested)
+            while (!stopProcessorsToken.IsCancellationRequested)
             {
                 try
                 {
                     var client = _reverseCallClientsCreator.Create(protocol);
 
-                    var connected = await client.Connect(eventProcessor.RegistrationRequest, cancellationToken).ConfigureAwait(false);
+                    var connected = await client.Connect(eventProcessor.RegistrationRequest, cancelConnectToken).ConfigureAwait(false);
                     if (!connected)
                     {
                         _logger.LogWarning("{Kind} {Identifier} failed to connect to the Runtime, retrying in 1s.", eventProcessor.Kind, eventProcessor.Identifier);
@@ -89,7 +91,7 @@ namespace Dolittle.SDK.Events.Processing
                     }
 
                     _logger.LogInformation("{Kind} {Identifier} registered with the Runtime, start handling requests", eventProcessor.Kind, eventProcessor.Identifier);
-                    await client.Handle(eventProcessor, cancellationToken).ConfigureAwait(false);
+                    await client.Handle(eventProcessor, stopProcessorsToken).ConfigureAwait(false);
                 }
                 catch (Exception exception)
                 {
