@@ -2,17 +2,18 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
-using Dolittle.SDK.Aggregates;
+using Dolittle.SDK.Aggregates.Builders;
 using Dolittle.SDK.DependencyInversion;
 using Dolittle.SDK.Embeddings;
 using Dolittle.SDK.EventHorizon;
 using Dolittle.SDK.Events;
-using Dolittle.SDK.Events.Store;
 using Dolittle.SDK.Events.Store.Builders;
 using Dolittle.SDK.Projections.Store.Builders;
 using Dolittle.SDK.Resources;
-using Dolittle.SDK.Tenancy.Client;
+using Dolittle.SDK.Tenancy;
 
 namespace Dolittle.SDK
 {
@@ -22,6 +23,11 @@ namespace Dolittle.SDK
     public interface IDolittleClient
     {
         /// <summary>
+        /// Gets a value indicating whether the Dolittle Client is connected.
+        /// </summary>
+        bool Connected { get; }
+
+        /// <summary>
         /// Gets the <see cref="IEventTypes" />.
         /// </summary>
         IEventTypes EventTypes { get; }
@@ -30,6 +36,11 @@ namespace Dolittle.SDK
         /// Gets the <see cref="IEventStoreBuilder" />.
         /// </summary>
         IEventStoreBuilder EventStore { get; }
+
+        /// <summary>
+        /// Gets the <see cref="IAggregatesBuilder"/>.
+        /// </summary>
+        IAggregatesBuilder Aggregates { get; }
 
         /// <summary>
         /// Gets the <see cref="IProjectionStoreBuilder" />.
@@ -47,9 +58,9 @@ namespace Dolittle.SDK
         IEventHorizons EventHorizons { get; }
 
         /// <summary>
-        /// Gets the <see cref="ITenants"/>.
+        /// Gets the <see cref="IEnumerable{T}"/> of <see cref="Tenant"/>.
         /// </summary>
-        ITenants Tenants { get; }
+        IEnumerable<Tenant> Tenants { get; }
 
         /// <summary>
         /// Gets the <see cref="IResourcesBuilder"/>.
@@ -57,35 +68,38 @@ namespace Dolittle.SDK
         IResourcesBuilder Resources { get; }
 
         /// <summary>
-        /// /// Sets the <see cref="IContainer" /> to use for inversion of control.
+        /// Gets the <see cref="ITenantScopedProviders"/>.
         /// </summary>
-        /// <param name="container">The <see cref="IContainer" /> to use for inversion of control.</param>
-        /// <returns>The client builder for continuation.</returns>
-        IDolittleClient WithContainer(IContainer container);
+        ITenantScopedProviders Services { get; }
 
         /// <summary>
-        /// Start the client.
+        /// Connects the <see cref="IDolittleClient"/>.
         /// </summary>
-        /// <returns>A <see cref="Task"/> that completes when the client has started. </returns>
-        Task Start();
+        /// <param name="configuration">The <see cref="DolittleClientConfiguration"/>.</param>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>A <see cref="Task"/>, that when resolved, returns the client for continuation.</returns>
+        Task<IDolittleClient> Connect(DolittleClientConfiguration configuration, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Gets the <see cref="IAggregateRootOperations{TAggregate}" /> for a new aggregate of the specified <typeparamref name="TAggregateRoot"/>.
+        /// Connects the <see cref="IDolittleClient"/>.
         /// </summary>
-        /// <param name="buildEventStore">The <see cref="Func{T, TResult}" /> for creating the <see cref="IEventStore" />.</param>
-        /// <typeparam name="TAggregateRoot">The <see cref="Type" /> of the <see cref="AggregateRoot" />.</typeparam>
-        /// <returns>The <see cref="IAggregateRootOperations{TAggregate}" />.</returns>
-        IAggregateRootOperations<TAggregateRoot> AggregateOf<TAggregateRoot>(Func<IEventStoreBuilder, IEventStore> buildEventStore)
-            where TAggregateRoot : AggregateRoot;
+        /// <param name="configureClient">The optional <see cref="Action{T}"/> callback for configuring the <see cref="DolittleClientConfiguration"/>.</param>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>A <see cref="Task"/>, that when resolved, returns the client for continuation.</returns>
+        Task<IDolittleClient> Connect(ConfigureDolittleClient configureClient, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Gets the <see cref="IAggregateRootOperations{TAggregate}" /> for a new aggregate of the specified <typeparamref name="TAggregateRoot"/>.
+        /// Connects the <see cref="IDolittleClient"/>.
         /// </summary>
-        /// <param name="eventSource">The <see cref="EventSourceId" />.</param>
-        /// <param name="buildEventStore">The <see cref="Func{T, TResult}" /> for creating the <see cref="IEventStore" />.</param>
-        /// <typeparam name="TAggregateRoot">The <see cref="Type" /> of the <see cref="AggregateRoot" />.</typeparam>
-        /// <returns>The <see cref="IAggregateRootOperations{TAggregate}" />.</returns>
-        IAggregateRootOperations<TAggregateRoot> AggregateOf<TAggregateRoot>(EventSourceId eventSource, Func<IEventStoreBuilder, IEventStore> buildEventStore)
-            where TAggregateRoot : AggregateRoot;
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>A <see cref="Task"/>, that when resolved, returns the client for continuation.</returns>
+        Task<IDolittleClient> Connect(CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Disconnects the <see cref="IDolittleClient"/>.
+        /// </summary>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+        Task Disconnect(CancellationToken cancellationToken = default);
     }
 }
